requirements.txt
SpeechRecognition
pipwin
pyaudio
gTTS
gtts
google-generativeai
python-dotenv
streamlit

-e .

